refactor coordinator
